requirements.txt
absl-py==1.4.0
ackermann-msgs==2.0.2
action-msgs==1.2.1
action-tutorials-interfaces==0.20.3
action-tutorials-py==0.20.3
actionlib-msgs==4.2.3
ament-cmake-test==1.3.8
ament-copyright==0.12.10
ament-cppcheck==0.12.10
ament-cpplint==0.12.10
ament-flake8==0.12.10
ament-index-python==1.4.0
ament-lint==0.12.10
ament-lint-cmake==0.12.10
ament-package==0.14.0
ament-pep257==0.12.10
ament-uncrustify==0.12.10
ament-xmllint==0.12.10
angles==1.15.0
array-record==0.5.0
astra-camera-msgs==1.0.1
asttokens==2.4.1
astunparse==1.6.3
bleach==6.1.0
bond==3.0.2
builtin-interfaces==1.2.1
cachetools==5.3.3
camera-calibration==3.0.4
cartographer-ros-msgs==2.0.9000
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
composition-interfaces==1.2.1
contourpy==1.2.0
control-msgs==4.4.0
controller-manager==2.40.0
controller-manager-msgs==2.40.0
cv-bridge==3.2.1
cycler==0.12.1
Cython==3.0.9
debugpy==1.8.1
decorator==5.1.1
demo-nodes-py==0.20.3
diagnostic-msgs==4.2.3
diagnostic-updater==3.1.2
dm-tree==0.1.8
domain-coordinator==0.10.0
dwb-msgs==1.1.13
einops==0.7.0
etils==1.7.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.1
examples-rclpy-minimal-action-client==0.15.1
examples-rclpy-minimal-action-server==0.15.1
examples-rclpy-minimal-client==0.15.1
examples-rclpy-minimal-publisher==0.15.1
examples-rclpy-minimal-service==0.15.1
examples-rclpy-minimal-subscriber==0.15.1
exceptiongroup==1.2.0
executing==2.0.1
exploration-behavior==0.0.0
find-object-2d==0.7.0
flatbuffers==24.3.7
fonttools==4.49.0
fsspec==2024.2.0
gast==0.4.0
gazebo-msgs==3.7.0
generate-parameter-library-py==0.3.7
geometry-msgs==4.2.3
gin-config==0.5.0
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
h5py==3.10.0
httplib2==0.22.0
idna==3.6
image-geometry==3.2.1
immutabledict==4.2.0
importlib_resources==6.3.0
install==1.3.5
interactive-markers==2.3.2
ipykernel==6.29.3
ipython==8.22.2
jax==0.4.25
jedi==0.19.1
joblib==1.3.2
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kaggle==1.6.6
keras==2.12.0
kiwisolver==1.4.5
laser-geometry==2.4.0
launch==1.0.5
launch-param-builder==0.1.1
launch-ros==0.19.7
launch-testing==1.0.5
launch-testing-ros==0.19.7
launch-xml==1.0.5
launch-yaml==1.0.5
libclang==16.0.6
lifecycle-msgs==1.2.1
limo-gazebosim==0.0.0
limo-msgs==0.0.0
logging-demo==0.20.3
lxml==5.1.0
map-msgs==2.1.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
message-filters==4.3.3
ml-dtypes==0.3.2
moveit-configs-utils==2.5.5
moveit-msgs==2.2.1
nav-2d-msgs==1.1.13
nav-msgs==4.2.3
nav2-common==1.1.13
nav2-msgs==1.1.13
nav2-simple-commander==1.0.0
nest-asyncio==1.6.0
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
object-detection==0.0.0
object-recognition-msgs==2.0.0
octomap-msgs==2.0.0
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
osrf-pycommon==2.0.2
packaging==24.0
pandas==2.2.1
parso==0.8.3
pcl-msgs==1.0.0
pendulum-msgs==0.20.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
portalocker==2.8.2
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-qt-binding==1.1.2
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.3
rcl-interfaces==1.2.1
rclpy==3.3.12
rcutils==5.1.5
regex==2023.12.25
report-generator==0.0.0
requests==2.31.0
requests-oauthlib==1.4.0
resource-retriever==3.1.1
rmw-dds-common==1.6.0
ros2action==0.18.9
ros2bag==0.15.9
ros2cli==0.18.9
ros2component==0.18.9
ros2controlcli==2.40.0
ros2doctor==0.18.9
ros2interface==0.18.9
ros2launch==0.19.7
ros2lifecycle==0.18.9
ros2multicast==0.18.9
ros2node==0.18.9
ros2param==0.18.9
ros2pkg==0.18.9
ros2run==0.18.9
ros2service==0.18.9
ros2topic==0.18.9
rosbag2-interfaces==0.15.9
rosbag2-py==0.15.9
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.4
rqt-bag-plugins==1.1.4
rqt-console==2.0.3
rqt-graph==1.3.0
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-tf-tree==1.0.4
rqt-topic==1.5.0
rsa==4.9
sacrebleu==2.4.1
scikit-learn==1.4.1.post1
scipy==1.12.0
scripts==3.7.0
sensor-msgs==4.2.3
sensor-msgs-py==4.2.3
sentencepiece==0.2.0
seqeval==1.2.2
shape-msgs==4.2.3
six==1.16.0
slam-toolbox==2.6.8
smclib==3.0.2
srdfdom==2.0.4
sros2==0.10.4
stack-data==0.6.3
statistics-msgs==1.2.1
std-msgs==4.2.3
std-srvs==4.2.3
stereo-msgs==4.2.3
tabulate==0.9.0
teleop-twist-keyboard==2.3.2
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-datasets==4.9.4
tensorflow-estimator==2.12.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-metadata==1.14.0
tensorflow-model-optimization==0.8.0
tensorflow-text==2.12.1
termcolor==2.4.0
text-unidecode==1.3
tf-keras==2.15.0
tf-models-official==2.15.0
tf-slim==1.1.0
tf-transformations==1.0.0
tf2-geometry-msgs==0.25.6
tf2-kdl==0.25.6
tf2-msgs==0.25.6
tf2-py==0.25.6
tf2-ros-py==0.25.6
tf2-tools==0.25.6
theora-image-transport==2.5.0
threadpoolctl==3.3.0
toml==0.10.2
topic-monitor==0.20.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
trajectory-msgs==4.2.3
turtlebot3-example==2.1.5
turtlebot3-msgs==2.2.3
turtlebot3-teleop==2.1.5
turtlesim==1.4.2
twist-mux-msgs==3.0.1
typing_extensions==4.10.0
tzdata==2024.1
unique-identifier-msgs==2.2.1
urdfdom-py==1.2.1
uritemplate==4.1.1
urllib3==2.2.1
visualization-msgs==4.2.3
voice-control==0.0.0
waypoint-follower==0.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.14.1
xacro==2.0.8
yolov8-object-detection==0.0.0
zipp==3.18.1
